+++ src/main/controller/prospectController.js
@@ -3,153 +3,15 @@ const TABLES = require('../variables/tables.js').TABLES;
 const HTTP = require('../variables/status.js').HTTP;
 const validator = require('../validator/prospectValidator');
 let PROSPECT_QUERY = require('../variables/prospect_sql.js').QUERY;
-const prospectIden_Service = require('../helper/prospect-record/prospect-identifier-service.js');
+const PROSPECT_HELPER = require('../helpers/prospect/prospect_helper.js');
+const PROSPECT_IDENTIFIER_HELPER = require('../helpers/prospect-record/prospect_identifier_helper.js');
+let IDENTIFIER = require('../variables/identifier.js').IDENTIFIER;
+const FIND_HELPER = require('../helpers/prospect-record/find_helper.js');
+const ADD_HELPER = require('../helpers/prospect-record/add_helper');
 let X_Auth = require('../variables/x-authorisation.json');
 let X_Auth_Find = require('../variables/x-auth-id-find.json');
 const X_Auth_Add = require('../variables/x-auth-add.json');
-let IDENTIFIER = require('../variables/identifier.js').IDENTIFIER;
-const findBy = require('../helper/prospect-record/find-service.js')
-const PROSPECT_UPDATE_COLS = ['brand_identifier', 'channel_identifier', 'first_name']
 
-/* checking if a Prospect present in the DB
-*/
-async function isProspectPresent(prospectId) {
-    const query = PROSPECT_QUERY.RECORD_COUNT
-        .replace('<tableName>', TABLES.PROSPECT)
-        .replace('<prospectId>', prospectId);
-
-    return (await db.getRecord(query))
-        .recordset[0].RECORD_COUNT !== 0 ? true : false;
-}
-
-async function getProspectWithSessionIdorIBID(SessionIdorIBID) {
-    const query = PROSPECT_QUERY.GET_PROSPECT_WITH_SESSION_ID_OR_IBID
-        .replace('<tableName>', TABLES.PROSPECT_IDENTIFIERS)
-        .replace('<identifier>', SessionIdorIBID);
-    var record = (await db.getRecord(query)).recordset[0]
-    if (record != null) {
-        return record.PROSPECT_ID;
-    } else {
-        return null;
-    }
-}
-
-async function getMaxProspectId() {
-    const query = PROSPECT_QUERY.GET_PROSPECT_ID
-        .replace('<tableName>', TABLES.PROSPECT);
-    return (await db.getRecord(query))
-        .recordset[0].MAXPROSPECTID;
-}
-
-async function getMaxProspectIdenId() {
-    const query = PROSPECT_QUERY.GET_PROSPECT_IDENTIFIER_ID
-        .replace('<tableName>', TABLES.PROSPECT_IDENTIFIERS);
-    return 'PID' + ((await db.getRecord(query))
-        .recordset[0].MAXPROSPECTIDENTIFIERID);
-}
-
-function getNextProspectIdenId(prospectIdenId) {
-    return (prospectIdenId == 'PIDnull') ? 'PID1'
-        : 'PID' + (parseInt(prospectIdenId.substring(3)) + 1);
-}
-
-/* getting the list of IdentifierType which is required to be archived
-return:
-('SessionId','EmailId','MobileNumber', ....)
-*/
-function getIdentifierTypeList(reqPayload) {
-    let updateFields = "'SessionId'";
-    for (let value of Object.values(reqPayload)) {
-        updateFields += (",'" + value.IdentifierType + "'");
-    }
-    return updateFields;
-}
-
-/* archiving the previous records by populating the ActiveTo column in Prospect_Identifiers table
-*/
-async function updateActiveTo(prospectId, reqPayload) {
-    const updateQuery = PROSPECT_QUERY.UPDATE_ACTIVETO
-        .replace('<tableName>', TABLES.PROSPECT_IDENTIFIERS)
-        .replace('<prospectId>', prospectId)
-        .replace('<identifierTypeList>', getIdentifierTypeList(reqPayload));
-    console.log(updateQuery)
-    await db.updateRecord(updateQuery);
-}
-
-function separateAddReqPayload(reqPayload) {
-    let prospect_payload = [];
-    let prospectIdentifier_payload = [];
-    for (let value of Object.values(reqPayload)) {
-        if (PROSPECT_UPDATE_COLS.includes(value.IdentifierType)) {
-            prospect_payload.push(value);
-        } else {
-            prospectIdentifier_payload.push(value);
-        }
-    }
-    return { prospect_payload, prospectIdentifier_payload };
-}
-
-/* getting list of values to be added in the Prospect_Identifiers table
-return: 
-('<prospectIdentifierId>',<prospectId>,'<identifier>','<identifierType>', '<activeFrom>')
-*/
-async function getInsertValues(prospectId, reqPayload) {
-    let insert_Val_list = '';
-    const lastItem = Object.values(reqPayload).pop();
-    let max_prospectIdenId = await getMaxProspectIdenId();
-
-    for (let value of Object.values(reqPayload)) {
-        let prospectIdenId = getNextProspectIdenId(max_prospectIdenId);
-        const insert_Val = PROSPECT_QUERY.INSERT_VALS
-            .replace('<prospectIdentifierId>', prospectIdenId)
-            .replace('<prospectId>', prospectId)
-            .replace('<identifier>', value.IdentifierValue)
-            .replace('<identifierType>', value.IdentifierType)
-            .replace('<activeFrom>', value.ActiveFrom)
-
-        insert_Val_list += insert_Val;
-        insert_Val_list += (value !== lastItem) ? ',' : '';
-        max_prospectIdenId = prospectIdenId;
-    }
-    return insert_Val_list;
-}
-
-/* getting list of columns and values to be updates in the Prospect_Identifiers table
-return: 
-<colName1>=<colValue1>,<colName2>=<colValue2>
-*/
-function getUpdateFields(payload) {
-    let update_fields = '';
-    const lastItem = Object.values(payload).pop();
-
-    for (let value of Object.values(payload)) {
-        update_fields += (value.IdentifierType + "='" + value.IdentifierValue + "'");
-        update_fields += (value !== lastItem) ? ',' : '';
-    }
-    console.log("\nupdate_fields: \n" + update_fields);
-    return update_fields;
-}
-
-/* adding Prospect Identifiers details to the already existing Prospect
-*/
-async function addProspectIdenRecord(prospectId, reqPayload) {
-    const insertQuery = PROSPECT_QUERY.ADD_PROSP_IDENTIFIER
-        .replace('<tableName>', TABLES.PROSPECT_IDENTIFIERS)
-        .replace('<insertVals>', await getInsertValues(prospectId, reqPayload));
-    console.log(insertQuery);
-    await db.insertRecord(insertQuery);
-}
-
-/* updating Prospect details to the already existing Prospect
-*/
-async function updateProspectRecord(prospectId, reqPayload) {
-    const updateQuery = PROSPECT_QUERY.UPDATE_PROSPECT
-        .replace('<tableName>', TABLES.PROSPECT)
-        .replace('<update_fields>', getUpdateFields(reqPayload))
-        .replace('<prospectId>', prospectId);
-    console.log("\nupdateQuery prospect: \n" + updateQuery);
-    await db.updateRecord(updateQuery);
-}
 
 // creating the prospect, if the customer id does not exist in the system
 async function createProspect(req, res) {
@@ -163,16 +25,16 @@ async function createProspect(req, res) {
     var ProspectIdfromDB
     var usertype = X_Auth[0].userType
     if (usertype === 'UNAUTH_CUSTOMER') {
-        ProspectIdfromDB = await prospectIden_Service.getProspectWithSessionId(X_Auth[0].sub)
+        ProspectIdfromDB = await PROSPECT_IDENTIFIER_HELPER.getProspectWithSessionId(X_Auth[0].sub)
     } else if (usertype === 'IB_CUSTOMER') {
-        ProspectIdfromDB = await prospectIden_Service.getProspectWithIBID(X_Auth[0].sub)
+        ProspectIdfromDB = await PROSPECT_IDENTIFIER_HELPER.getProspectWithIBID(X_Auth[0].sub)
     } else {
         res.status(HTTP.NOT_FOUND.code)
             .json({ message: `User_Type: ${usertype} is invalid.` });
         return
     }
     if (ProspectIdfromDB == null) {
-        var prevProspectId = await getMaxProspectId();
+        var prevProspectId = await PROSPECT_HELPER.getMaxProspectId();
         var newProspectId = prevProspectId == null ? 10000000 : (parseInt(prevProspectId) + 1);
         const insertProspectQuery = PROSPECT_QUERY.INSERT_PROSPECT
             .replace('<tableName>', TABLES.PROSPECT)
@@ -184,8 +46,8 @@ async function createProspect(req, res) {
 
         const prospectInsertResult = await db.insertRecord(insertProspectQuery);
 
-        var prevProspectIdentifierId = await getMaxProspectIdenId();
-        var newProspectIdentifierId = getNextProspectIdenId(prevProspectIdentifierId)
+        var prevProspectIdentifierId = await PROSPECT_IDENTIFIER_HELPER.getMaxProspectIdenId();
+        var newProspectIdentifierId = PROSPECT_IDENTIFIER_HELPER.getNextProspectIdenId(prevProspectIdentifierId)
 
         var usertype = X_Auth[0].userType
         if (usertype === 'UNAUTH_CUSTOMER') {
@@ -229,24 +91,24 @@ async function addProspectById(req, res) {
     }
 
     if (X_Auth_Add.userType === "UNAUTH_CUSTOMER") {
-        const dbProspectId = await prospectIden_Service.getProspectWithSessionId(X_Auth_Add.sub);
+        const dbProspectId = await PROSPECT_IDENTIFIER_HELPER.getProspectWithSessionId(X_Auth_Add.sub);
 
         if (dbProspectId == null || dbProspectId != reqParams.ProspectId) {
             res.status(HTTP.NOT_FOUND.code)
                 .json({ error: `ProspectId: ${dbProspectId}, is not associated with Auth SessionId: ${X_Auth_Add.sub}` });
         } else {
-            let { prospect_payload, prospectIdentifier_payload } = separateAddReqPayload(reqPayload);
+            let { prospect_payload, prospectIdentifier_payload } = ADD_HELPER.separateAddReqPayload(reqPayload);
 
             if (Object.keys(prospect_payload).length !== 0) {
-                await updateProspectRecord(dbProspectId, prospect_payload);
+                await ADD_HELPER.updateProspectRecord(dbProspectId, prospect_payload);
             } else {
-                console.log("Payload doesnot contain tbl_prospect record to be updated");
+                console.log("\nPayload doesnot contain tbl_prospect record to be updated");
             }
             if (Object.keys(prospectIdentifier_payload).length !== 0) {
-                await updateActiveTo(dbProspectId, prospectIdentifier_payload);
-                await addProspectIdenRecord(dbProspectId, prospectIdentifier_payload);
+                await ADD_HELPER.updateActiveTo(dbProspectId, prospectIdentifier_payload);
+                await ADD_HELPER.addProspectIdenRecord(dbProspectId, prospectIdentifier_payload);
             } else {
-                console.log("Payload doesnot contain tbl_prospect_identifier record to be updated");
+                console.log("\nPayload doesnot contain tbl_prospect_identifier record to be updated");
             }
 
             res.status(HTTP.OK.code)
@@ -270,7 +132,7 @@ async function addProspect(req, res) {
     }
 
     if (X_Auth_Add.userType === "UNAUTH_CUSTOMER") {
-        const dbProspectId = await prospectIden_Service.getProspectWithSessionId(X_Auth_Add.sub);
+        const dbProspectId = await PROSPECT_IDENTIFIER_HELPER.getProspectWithSessionId(X_Auth_Add.sub);
 
         if (dbProspectId == null) {
             res.status(HTTP.NOT_FOUND.code)
@@ -281,13 +143,13 @@ async function addProspect(req, res) {
             if (Object.keys(prospect_payload).length !== 0) {
                 await updateProspectRecord(dbProspectId, prospect_payload);
             } else {
-                console.log("Payload doesnot contain tbl_prospect record to be updated");
+                console.log("\nPayload doesnot contain tbl_prospect record to be updated");
             }
             if (Object.keys(prospectIdentifier_payload).length !== 0) {
                 await updateActiveTo(dbProspectId, prospectIdentifier_payload);
                 await addProspectIdenRecord(dbProspectId, prospectIdentifier_payload);
             } else {
-                console.log("Payload doesnot contain tbl_prospect_identifier record to be updated");
+                console.log("\nPayload doesnot contain tbl_prospect_identifier record to be updated");
             }
 
             res.status(HTTP.OK.code)
@@ -312,7 +174,7 @@ async function findProspect(req, res) {
             .send(error.details);
     }
 
-    const result = await findBy.findProspect(req);
+    const result = await FIND_HELPER.findProspect(req);
     console.log(result.error);
 
     if (result.error == null) {
@@ -337,7 +199,7 @@ async function findProspectById(req, res) {
             .send(error.details);
     }
 
-    const result = await findBy.findProspectById(req);
+    const result = await FIND_HELPER.findProspectById(req);
     console.log(result.error);
 
     if (result.error == null) {
