+++ src/main/helpers/prospect-record/prospect_identifier_helper.js
@@ -2,6 +2,19 @@ let PROSPECT_QUERY = require('../../variables/prospect_sql.js').QUERY;
 const TABLES = require('../../variables/tables.js').TABLES;
 const db = require('../../utils/azureSql.js');
 
+
+async function getMaxProspectIdenId() {
+    const query = PROSPECT_QUERY.GET_PROSPECT_IDENTIFIER_ID
+        .replace('<tableName>', TABLES.PROSPECT_IDENTIFIERS);
+    return 'PID' + ((await db.getRecord(query))
+        .recordset[0].MAXPROSPECTIDENTIFIERID);
+}
+
+function getNextProspectIdenId(prospectIdenId) {
+    return (prospectIdenId == 'PIDnull') ? 'PID1'
+        : 'PID' + (parseInt(prospectIdenId.substring(3)) + 1);
+}
+
 async function getProspectWithSessionId(SessionId) {
     const query = PROSPECT_QUERY.GET_PROSPECT_WITH_SESSION_ID
         .replace('<tableName>', TABLES.PROSPECT_IDENTIFIERS)
@@ -32,4 +45,4 @@ async function getProspectWithIBID(IBID) {
 }
 
 // exporting modules, to be used in the other .js files
-module.exports = { getProspectWithSessionId, getProspectWithIBID };
+module.exports = { getMaxProspectIdenId, getNextProspectIdenId, getProspectWithSessionId, getProspectWithIBID };
